requirements.txt
fastapi>=0.68.0
uvicorn>=0.23.0
openai>=1.12.0
chromadb>=0.4.22
langchain>=0.1.13
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
langgraph>=0.0.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
pandas>=2.0.0
openpyxl>=3.0.0
python-docx>=0.8.0
PyPDF2>=3.0.0
